Move tabToIcon to where it's used
